refactor checkout.go and introduce internal package
