+++ checkout.go
@@ -12,6 +12,7 @@ import (
 	"time"
 
 	"github.com/checkout/checkout-sdk-go/common"
+	"github.com/checkout/checkout-sdk-go/internal/utils"
 )
 
 // ClientVersion ...
@@ -71,7 +72,7 @@ type Config struct {
 
 // DefaultConfig ...
 var DefaultConfig = Config{
-	URI: String(sandboxURI),
+	URI: utils.String(sandboxURI),
 }
 
 const (
@@ -99,7 +100,7 @@ func Create(secretKey string, publicKey *string) (*Config, error) {
 	}
 
 	if config.MaxNetworkRetries == nil {
-		config.MaxNetworkRetries = Int64(DefaultMaxNetworkRetries)
+		config.MaxNetworkRetries = utils.Int64(DefaultMaxNetworkRetries)
 	}
 
 	if publicKey == nil {
@@ -108,8 +109,8 @@ func Create(secretKey string, publicKey *string) (*Config, error) {
 
 	if !isSandbox {
 		publicKeyMatch := regexp.MustCompile(common.LivePublicKeyRegex)
-		if publicKeyMatch.MatchString(StringValue(publicKey)) {
-			config.PublicKey = StringValue(publicKey)
+		if publicKeyMatch.MatchString(utils.StringValue(publicKey)) {
+			config.PublicKey = utils.StringValue(publicKey)
 			return &config, nil
 		}
 		return nil, &common.Error{
@@ -117,8 +118,8 @@ func Create(secretKey string, publicKey *string) (*Config, error) {
 		}
 	}
 	publicKeyMatch := regexp.MustCompile(common.SandboxPublicKeyRegex)
-	if publicKeyMatch.MatchString(StringValue(publicKey)) {
-		config.PublicKey = StringValue(publicKey)
+	if publicKeyMatch.MatchString(utils.StringValue(publicKey)) {
+		config.PublicKey = utils.StringValue(publicKey)
 		return &config, nil
 	}
 	return nil, &common.Error{
@@ -131,12 +132,12 @@ func create(secretKey string) (Config, bool) {
 	liveSecretKeyMatch := regexp.MustCompile(common.LiveSecretKeyRegex)
 	if liveSecretKeyMatch.MatchString(secretKey) {
 		return Config{
-			URI:       String(productionURI),
+			URI:       utils.String(productionURI),
 			SecretKey: secretKey,
 		}, false
 	}
 	return Config{
-		URI:       String(sandboxURI),
+		URI:       utils.String(sandboxURI),
 		SecretKey: secretKey,
 	}, true
 }
@@ -251,87 +252,4 @@ type HTTPClient interface {
 	Delete(path string) (*StatusResponse, error)
 	Upload(path, boundary string, body *bytes.Buffer) (*StatusResponse, error)
 	Download(path string) (*StatusResponse, error)
-}
-
-// Int64 returns a pointer to the int64 value passed in.
-func Int64(v int64) *int64 {
-	return &v
-}
-
-// Int64Value returns the value of the int64 pointer passed in or
-// 0 if the pointer is nil.
-func Int64Value(v *int64) int64 {
-	if v != nil {
-		return *v
-	}
-	return 0
-}
-
-// String returns a pointer to the string value passed in.
-func String(v string) *string {
-	return &v
-}
-
-// StringValue returns the value of the string pointer passed in or
-// "" if the pointer is nil.
-func StringValue(v *string) string {
-	if v != nil {
-		return *v
-	}
-	return ""
-}
-
-// StringSlice returns a slice of string pointers given a slice of strings.
-func StringSlice(v []string) []*string {
-	out := make([]*string, len(v))
-	for i := range v {
-		out[i] = &v[i]
-	}
-	return out
-}
-
-// Bool returns a pointer to the bool value passed in.
-func Bool(v bool) *bool {
-	return &v
-}
-
-// BoolValue returns the value of the bool pointer passed in or
-// false if the pointer is nil.
-func BoolValue(v *bool) bool {
-	if v != nil {
-		return *v
-	}
-	return false
-}
-
-// BoolSlice returns a slice of bool pointers given a slice of bools.
-func BoolSlice(v []bool) []*bool {
-	out := make([]*bool, len(v))
-	for i := range v {
-		out[i] = &v[i]
-	}
-	return out
-}
-
-// Float64 returns a pointer to the float64 value passed in.
-func Float64(v float64) *float64 {
-	return &v
-}
-
-// Float64Value returns the value of the float64 pointer passed in or
-// 0 if the pointer is nil.
-func Float64Value(v *float64) float64 {
-	if v != nil {
-		return *v
-	}
-	return 0
-}
-
-// Float64Slice returns a slice of float64 pointers given a slice of float64s.
-func Float64Slice(v []float64) []*float64 {
-	out := make([]*float64, len(v))
-	for i := range v {
-		out[i] = &v[i]
-	}
-	return out
 }
